+++ src/collective/folderishtypes/browser/listing_portlet.py
@@ -6,7 +6,6 @@ from plone.app.portlets.portlets import base
 from plone.portlets.interfaces import IPortletDataProvider
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope.component import getMultiAdapter
-from zope.formlib import form
 from zope.interface import implementer
 
 
@@ -26,7 +25,6 @@ class Assignment(base.Assignment):
 
 
 class AddForm(base.NullAddForm):
-    form_fields = form.Fields(IListingPortlet)
     label = _(u"listingportlet_label_add", u"Add portlet for folderish types")
     description = _(
         u"listingportlet_help_add",